mise à jour de l'interface utilisateur, Makefile fonctionnel

+++ include/operations.h
@@ -4,7 +4,7 @@
 
 unsigned int pgcd(int a, int b);
 unsigned int ppcm(int a, int b);
-unsigned int factoreil(unsigned int);
+unsigned long factoreil(unsigned int);
 int multiplication(int a, int b);
 float soustraction(float a, float b);
 int somme(int a, int b);
